+++ src/main/java/ru/mail/polis/ads/part4.java
@@ -1,71 +1,78 @@
 package ru.mail.polis.ads;
 
-
+import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.lang.reflect.Array;
-import java.util.LinkedList;
+import java.util.ArrayList;
 import java.io.BufferedReader;
 
 public class part4 {
     public static void main(String[] arg) throws IOException {
-        LinkedList arr = new LinkedList();
-        QuCust a = new QuCust(arr);
+        ArrayList arr = new ArrayList();
+        StackCust a = new StackCust(arr);
         BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
         String del = " ";
         String[] subStr;
         while(true){
-            subStr = null;
-            String com = reader.readLine();
-            subStr = com.split(del);
-            switch (subStr[0]){
-                case "push": {
-                    a.push(Integer.parseInt((subStr[1])));
-                    break;
-                }
-                case "pop": {
-                    a.pop();
-                    break;
-                }
-                case "front": {
-                    a.front();
-                    break;
-                }
-                case "size": {
-                    a.size();
-                    break;
-                }
-                case "clear": {
-                    a.clear();
-                    break;
-                }
-                case "exit": {
-                    a.exit();
-                    break;
-                }
+        subStr = null;
+        String com = reader.readLine();
+        subStr = com.split(del);
+        switch (subStr[0]){
+            case "push": {
+                a.push(Integer.parseInt((subStr[1])));
+                break;
+            }
+            case "pop": {
+                a.pop();
+                break;
+            }
+            case "back": {
+                a.back();
+                break;
+            }
+            case "size": {
+                a.size();
+                break;
+            }
+            case "clear": {
+                a.clear();
+                break;
+            }
+            case "exit": {
+                a.exit();
+                break;
             }
         }
+        }
     }
-    public static class QuCust {
-        private LinkedList arr;
-        public QuCust(LinkedList arr1){
+    public static class StackCust {
+        private final ArrayList arr;
+        public StackCust(ArrayList arr1){
             this.arr = arr1;
         }
         public void push(int n){
-            arr.add(arr.size(),n);
+            arr.add(n);
             System.out.println("ok");
         }
         public void pop(){
-            int a = (int) arr.get(0);
-            arr.remove(0);
-            System.out.println(a);
+            if(arr.size() == 0){
+                System.out.println("error");
+            }else{
+                int a = (int) arr.get(arr.size()-1);
+                arr.remove(arr.size()-1);
+                System.out.println(a);
+             }
         }
-        public void front(){
-            int a = (int) arr.get(0);
-            System.out.println(a);
+        public void back(){
+            if(arr.size() == 0){
+                System.out.println("error");
+            }else{
+                int a = (int) arr.get(arr.size()-1);
+                System.out.println(a);
+            }
         }
         public void size(){
-            System.out.println(arr.size());;
+            System.out.println(arr.size());
         }
         public void clear(){
             arr.clear();
